x64 corrections/updates: vcproj files updated with the X64 preprocessor directive; Vertex2f padded out to a multiple of a 64-bit word; #ifdef WIN32 transformed to #ifdef _WINDOWS; WIN32 transformed to WIN64 in x64 build targets; some instances of unsigned long transformed to Uint32;

+++ include/Zeni/Color.h
@@ -45,6 +45,8 @@
 #ifndef ZENI_COLOR_H
 #define ZENI_COLOR_H
 
+#include <SDL/SDL_stdinc.h>
+
 namespace Zeni {
 
   class Color {
@@ -64,8 +66,8 @@ namespace Zeni {
     inline unsigned char g_ub() const; ///< Get the green channel [0x00, 0xFF]
     inline unsigned char b_ub() const; ///< Get the blue channel [0x00, 0xFF]
 
-    inline unsigned long get_rgba() const; ///< Get a Uint32 representation of 0xRRGGBBAA
-    inline unsigned long get_argb() const; ///< Get a Uint32 representation of 0xAARRGGBB
+    inline Uint32 get_rgba() const; ///< Get a Uint32 representation of 0xRRGGBBAA
+    inline Uint32 get_argb() const; ///< Get a Uint32 representation of 0xAARRGGBB
 
     Color interpolate_to(const float &rhs_part, const Color &rhs) const; ///< Get a color that is inbetween this color and another color.
