+++ include/Zeni/Vertex3f.h
@@ -89,13 +89,13 @@ namespace Zeni {
     Vertex3f_Color(const Point3f &position, const Vector3f &normal, 
       const long &argb);
     Vertex3f_Color(const Point3f &position, const Color &color);
-    Vertex3f_Color(const Point3f &position, const long &argb);
+    Vertex3f_Color(const Point3f &position, const Uint32 &argb);
 
     Vertex3f * interpolate_to(const float &rhs_part, const Vertex3f_Color &rhs) const; ///< Get a Vertex3f_Color between two vertices; rhs must be a Vertex3f_Color
 
-    inline const unsigned long & get_color() const; ///< Get the current Color
+    inline const Uint32 & get_color() const; ///< Get the current Color
     inline void set_color(const Color &color); ///< Set the current Color
-    inline void set_color(const unsigned long &argb); ///< Set the current Color
+    inline void set_color(const Uint32 &argb); ///< Set the current Color
 
     // Begin rendering functions
 
@@ -111,7 +111,7 @@ namespace Zeni {
 #endif
 
   private:
-    unsigned long m_argb;
+    Uint32 m_argb;
   };
 
   class Vertex3f_Texture : public Renderable, public Vertex3f {
